Запись в БД - CREATE
